Changed behavior of (de)activating an item within the change form:
Instead of changing all descendants' active status to the current item's, it will only change the descendants' active status if the item is False.

As it makes sense to have an item active, but its children inactive, it doesn't make sense that an item is inactive, but its descendants are active.

This doesn't change the activate/deactivate admin actions. They will always  affect an item and its descendants.