Fix 'Models aren't loaded yet' warning on import
categories.registration._process_registry was being called in
categories/__init__.py, but since Django 1.9 it hasn't been possible to
perform operations with models until the app registry is fully
loaded. Currently the `AppRegistryNotReady` exception is being caught
and printed, which means it is never actually executed on load.

Since this code isn't currently doing anything (other than emitting a
print() of a warning), I've removed it.